requirements.txt
bidict==0.23.1
blinker==1.9.0
cachelib==0.13.0
click==8.1.8
dnspython==2.7.0
eventlet==0.39.1
Flask==3.1.0
flask-cors==5.0.1
Flask-Session==0.8.0
Flask-SocketIO==5.5.1
greenlet==3.1.1
h11==0.14.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
msgspec==0.19.0
python-engineio==4.11.2
python-socketio==5.12.1
redis==5.2.1
simple-websocket==1.1.0
Werkzeug==3.1.3
wsproto==1.2.0
antlr4-python3-runtime==4.13.2
